Source revisited
In this branch, I made some changes to the drawing logic. This Android/Java implementation works now exactly the same as the Javascript variant https://github.com/stephanbrunker/signature_pad/tree/stephanbrunker-patch-1 , which I have also revisited. The major change was the handling of the mPoints array. I changed this into an ArrayDeque, because it is more a queue of points to process than a List. I also changed the handling of the ACTION_DOWN / ACTION_MOVE / ACTION_UP events. The DOWN event has different coordinates than the first MOVE event, but the last MOVE event and the UP event are on the same position. To reflect that, I changed the handling of the UP event, to decide: 
- If 4 points are in the queue, double the last point to draw the bezier to the last position
- 3 points is not possible, because then the first point would be doubled to begin bezier drawing
- If 2 points are registered, draw a line
- if only one point is registered, draw a point (the previous version didn't lik